refactor: Remove Ankr as a default provider (#305)
`rpc.ankr.com` recently dropped its IPv6 connectivity and will need
according to its support team a month to fix it (see this forum
[post](https://forum.dfinity.org/t/cketh-a-canister-issued-ether-twin-token-on-the-ic/22819/204))
and is therefore unfit to be selected as a default provider in case the
caller did not manually specify the requested providers. This PR
replaces it with BlockPi (`ethereum.blockpi.network`), which has been so
far working well for the [ckETH
minter](https://dashboard.internetcomputer.org/canister/sv3dd-oaaaa-aaaar-qacoa-cai).