fix: always deserialize `Block::totalDifficulty` to `None` (#312)
As explained in #311, the field `totalDifficulty` was removed from the
official Ethereum JSON RPC Block schema in
https://github.com/ethereum/execution-apis/pull/570, leading to
inconsistent answers between providers when calling the EVM RPC canister
with `eth_getBlockByNumber`. This seems to affect all blocks and not
only post-merge blocks.

As a workaround, the EVM RPC canister no longer deserializes the field
`totalDifficulty` in the provider's JSON response and set the Candid
returned value for `evm_rpc_types::Block::totalDifficulty` always to
`None` to be backwards-compatible.